split graph/index into make, query and invalidate
