+++ src/graph/util.js
@@ -1,0 +1,37 @@
+import t from 'tcomb';
+import pick from 'lodash/pick';
+
+// given a flat `AA` object, e.g.
+//
+//   { a: 3, foo: 'asd', bar: true }
+//
+// and an (possibly nested) `A` description, e.g.
+//
+//   [['a'], ['foo', 'bar']]
+//
+// returns an (possibly nested array containing) object(s) with the values
+// corresponding to the given A keys, e.g.
+//
+//   [{ a: 3 }, { foo: 'asd', bar: true }]
+//
+function pickA(AA, A) {
+  if (t.list(t.String).is(A)) { // ['foo', 'bar']
+    return pick(AA, A);
+  } else {                      // [['foo', 'bar'], ['foo']]
+    return A.map(a => pickA(AA, a));
+  }
+}
+
+// given an array of `Ps`, aka "query ids"
+// and an object of arguments `A`
+// returns the same information rearranged to match the
+// lower level `apply` signature expectations
+export function queriesAndArgs(graph, Ps, A) {
+  const queries = Ps.reduce((qs, P) => Object.assign(qs, {
+    [P]: graph[P].cachedFetch || graph[P].fetch // non-atoms are not cached, we return the naked fetch
+  }), {});
+  const args = Ps.reduce((argz, P) => Object.assign(argz, {
+    [P]: pickA(A, graph[P].A)
+  }), {});
+  return { queries, args };
+}
